+++ pages/api/_constants.ts
@@ -3,13 +3,14 @@
 import {keypairIdentity} from "@metaplex-foundation/js";
 import {Keypair} from "@solana/web3.js";
 import * as bs58 from "bs58";
-import {getMetaplex, getSolanaConnection} from "./_util";
 import {AuthorizationFailureResponse} from "./_responses";
+import {getConnection} from "../../utils/get-connection";
+import {getMetaplex} from "../../utils/get-metaplex";
 
 // init'd objects to be used in API (and not Pages, which should use contexts)
 // assumes that APIs using the objects will use default networks (production -> mainnet, otherwise devnet)
 // if an API needs to use a specific network, it should call the getters itself
-export const connection = getSolanaConnection();
+export const connection = getConnection();
 export const metaplex = getMetaplex();
 
 export const adminWallet = Keypair.fromSecretKey(bs58.decode(process.env.SOLANA_PRIVATE_KEY!));
